Refactor AdminView component: replace inline table logic with InviteTable component for improved readability and maintainability
